Use short version of flags; fix url error handling
Modified usage statement in code and README.md to show short
versions of command line flags (-h and -d) and not the longer
"--debug".

Also added logic and a test for URL in the config file that
is missing a trailing slash.  A slash is added if one is
missing.

Also fixed the error handling for a response; was attempting
to print 'err' when the err was non-null.  Also a more
readable version of the status code should have been printed.